+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+  - Fix inverted constants in survey transform.
 
 ### 1.0.0 2017-02-16
   - Initial release